refactor: re-organize components in their specific directories and get rid of index.js files
