--- frontend/src/views/components/index.js
@@ -1,6 +1,0 @@
-import AppNavbar from './navbar'
-import CMSOMSCard from './cmsOMSCard'
-import ResponsivePlot from './responsivePlot'
-import Table from './table'
-
-export { AppNavbar, CMSOMSCard, ResponsivePlot, Table }
