+++ frontend/src/views/components/overview.jsx
@@ -6,9 +6,9 @@ import Col from 'react-bootstrap/Col'
 import Card from 'react-bootstrap/Card'
 import { toast } from 'react-toastify'
 
-import { ResponsivePlot } from './components'
-import API from '../services/api'
-import { groupBySplitME } from '../utils/ops'
+import API from '../../services/api'
+import { groupBySplitME } from '../../utils/ops'
+import ResponsivePlot from './responsivePlot'
 
 const Overview = () => {
   const [totalFiles, setTotalFiles] = useState(0)
