--- frontend/src/views/dataExplorer/index.js
@@ -1,19 +1,0 @@
-import FileIndex from './fileIndex'
-import Runs from './runs'
-import Run from './run'
-import Lumisections from './lumisections'
-import Lumisection from './lumisection'
-import Histograms1D from './histograms1d'
-import Histograms2D from './histograms2d'
-import Histogram from './histogram'
-
-export {
-  FileIndex,
-  Runs,
-  Run,
-  Lumisections,
-  Lumisection,
-  Histograms1D,
-  Histograms2D,
-  Histogram,
-}
