+++ frontend/src/views/dataExplorer/run.jsx
@@ -8,8 +8,9 @@ import Card from 'react-bootstrap/Card'
 import { toast } from 'react-toastify'
 
 import API from '../../services/api'
-import { CMSOMSCard, Table } from '../components'
 import { getNextToken } from '../../utils/sanitizer'
+import CMSOMSCard from '../components/cmsOMSCard'
+import Table from '../components/table'
 
 const Run = () => {
   const { datasetId, runNumber } = useParams()
