--- frontend/src/views/machineLearning/index.js
@@ -1,4 +1,0 @@
-import Predictions from './predictions'
-import JsonPortal from './jsonPortal'
-
-export { Predictions, JsonPortal }
